Core native compiles / green tests
