+++ modules/core/src-native/smithy4s/TimestampPlatform.scala
@@ -1,0 +1,7 @@
+package smithy4s
+
+import java.time.Instant
+import java.time.ZoneOffset
+import java.time.OffsetDateTime
+
+private[smithy4s] trait TimestampPlatform
