+++ modules/core/test/src-js/smithy4s/Platform.scala
@@ -3,4 +3,5 @@ package smithy4s
 object Platform {
   def isJS = true
   def isJVM = false
+  def isNative = false
 }
