+++ modules/core/test/src-native/smithy4s/Platform.scala
@@ -2,6 +2,6 @@ package smithy4s
 
 object Platform {
   def isJS = false
-  def isJVM = true
-  def isNative = false
+  def isJVM = false
+  def isNative = true
 }
