+++ modules/core/test/src/smithy4s/http/internals/MetadataSpec.scala
@@ -119,7 +119,7 @@ class MetadataSpec() extends FunSuite {
     "length required to be >= 1 and <= 10, but was 11"
 
   val constraintMessage2 =
-    if (Platform.isJVM) "Input must be >= 1.0 and <= 10.0, but was 11.0"
+    if (!Platform.isJS) "Input must be >= 1.0 and <= 10.0, but was 11.0"
     else "Input must be >= 1 and <= 10, but was 11"
 
   // ///////////////////////////////////////////////////////////
