+++ project/Smithy4sPlugin.scala
@@ -350,6 +350,7 @@ object Smithy4sPlugin extends AutoPlugin {
   }
 
   lazy val jvmDimSettings = simpleJVMLayout
+  lazy val nativeDimSettings = simpleNativeLayout
 
   lazy val simpleJSLayout = simpleLayout(JSPlatform)
   lazy val simpleJVMLayout = simpleLayout(JVMPlatform)
