+++ project/plugins.sbt
@@ -11,6 +11,7 @@ addSbtPlugin("com.eed3si9n"         % "sbt-buildinfo"                 % "0.11.0"
 addSbtPlugin("com.eed3si9n"         % "sbt-projectmatrix"             % "0.9.0")
 addSbtPlugin("pl.project13.scala"   % "sbt-jmh"                       % "0.4.3")
 addSbtPlugin("de.heikoseeberger"    % "sbt-header"                    % "5.7.0")
+addSbtPlugin("org.scala-native"     % "sbt-scala-native"              % "0.4.5")
 
 libraryDependencies ++= Seq("com.lihaoyi" %% "os-lib" % "0.8.1")
